DOI is not a valid/resolvable DOI
The DOI to the Gibb et al preprint is not a known or resolvable DOI because of the `v1` appended to the end of the DOI. This PR corrects the DOI so that is resolves to the biorxiv preprint. Additionally, the DOI has been turned into a hyperlink to facilitate accessing the preprint.